Add files via upload
All the programs are part of a sequential programming course in Spanish, for more information you can look for it in YouTube, the name of the channel is  pildorasinformaticas. I am adding the code here as a help for you, also I recommend to do it by hand and practice with additional problems related to the topics. That way you would reinforce your knowledge and it would be easy to remember after.